ci(linux): unblock ubuntu build + test job
Three real failures only surface on the linux/gcc lane (not macos/clang).
Verified locally via docker (rust:1.94-bookworm image) running the same
`cargo build --workspace --all-targets` and `cargo test --workspace
--no-fail-fast` commands `.github/workflows/ci.yml` runs.

* library/std/sync/runtime/atomic.c: gcc rejects
  `atomic_fetch_and/or` on `atomic_bool` (C11 defines those operations
  for integer atomic types only — Clang on macOS accepts the bool
  form as a non-standard extension, gcc on linux refuses):

      error: operand type `_Atomic _Bool *` is incompatible with
             argument 1 of `__atomic_fetch_and`

  Replaced both bool ops with a portable compare-exchange loop —
  same semantics, both compilers happy. Unblocks the ubuntu workspace
  build via the riven_repl build script (cc::Build).

* src/riven_cli/src/explain.rs: 9 codes (E1100, E1101, E1102, E1110,
  E1112, E1115, E1116, E1117, E1118) were in the central REGISTRY but
  missing from the EXPLAINS table — `cargo test
  every_registered_code_has_embedded_markdown` failed on linux even
  though the `docs/errors/<code>.md` files all existed. Added the 9
  `include_str!` lines.

* src/riven_ide/src/line_index.rs: `position_of` / `byte_offset_of`
  sliced `&self.source[line_start..end]` with byte offsets that
  weren't guaranteed to land on UTF-8 char boundaries. Sources
  containing `─` / `…` / emoji (e.g. `sample_program.rvn`'s header
  banner) panicked with "byte index N is not a char boundary; it is
  inside `─` (bytes M..N+3)". Added a `floor_char_boundary` helper
  (stable equivalent of the unstable
  `str::floor_char_boundary`) and routed both bounds through it.
  Pins `analysis_of_sample_program` and the v1 STATUS.md follow-up
  noted at #15.

Drive-by from the same docker run:

* src/riven_ide/src/{document_symbols,rename}.rs: removed three
  unnecessary `as usize` casts on `Span::end` (now already `usize`) +
  unnecessary parens that clippy flagged on the workspace lint.
  Applied via `cargo clippy --fix -p riven_ide`.

* src/{rivenc/tests/installed_binary,riven_cli/tests/installed_pkg_manager}.rs:
  deleted dead `runtime_c_src` helpers (zero callers). Cited in the
  quality review Wave 1 #3 and surfaced again as a clippy
  `dead_code` warning on linux.

Verification (local macOS aarch64 + docker linux):

  docker run riven-ci cargo build --workspace --all-targets   # green
  docker run riven-ci cargo test  --workspace --no-fail-fast  # 1643 pass / 0 fail
  cargo +1.91   build --workspace                              # green (MSRV macOS)
  cargo +stable test --release -p riven_core --tests           # 1066 pass / 0 fail
  cargo +stable fmt  --all -- --check                          # green

+++ src/riven_ide/src/document_symbols.rs
@@ -83,7 +83,7 @@ fn item_in_source(item: &HirItem, source_len: usize) -> bool {
 }
 
 fn span_within(span: &Span, source_len: usize) -> bool {
-    (span.end as usize) <= source_len
+    span.end <= source_len
 }
 
 fn item_top_span(item: &HirItem) -> &Span {
